# docs/swagger.yaml
basePath: /
definitions:
...
  model.Board:
    properties:
      description:
        type: string
      id:
        type: integer
      profileID:
        type: integer
      title:
        type: string
    type: object
...
  model.Todo:
    properties:
      boardID:
        description: "BoardID \tid of board which has this todo."
        type: integer
      description:
        type: string
      finished:
        type: boolean
      id:
        type: integer
      title:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  title: Swagger
  version: "1.0"
paths:
  /board/{boardid}/update:
    put:
      consumes:
      - application/json
      description: given new data and id, update board
      parameters:
      - description: ID of the to be updated board
        in: path
        name: boardid
        required: true
        type: integer
      - description: Update content
        in: body
        name: UpdateContent
        required: true
        schema:
          $ref: '#/definitions/model.Board'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: get board updated with new data
      tags:
      - board-manager-apis
  /board/{uid}/allBoard:
    get:
      consumes:
      - application/json
      description: gets all board belong to UserID
      parameters:
      - description: User ID
        in: path
        name: uid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: gets all board belong to UserID
      tags:
      - board-manager-apis
  /board/{uid}/create:
    post:
      consumes:
      - application/json
      description: create a new board with given model
      parameters:
      - description: Owner of this board
        in: path
        name: uid
        required: true
        type: integer
      - description: Board information
        in: body
        name: BoardInfo
        required: true
        schema:
          $ref: '#/definitions/model.Board'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: create new Board with given model
      tags:
      - board-manager-apis
  /board/delete/{boardid}:
    delete:
      consumes:
      - application/json
      description: Board with boardID will be deleted
      parameters:
      - description: ID of the to be deleted board
        in: path
        name: boardid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: Board with boardID will be deleted
      tags:
      - board-manager-apis
  /sys/allBoard:
    get:
      consumes:
      - application/json
      description: get all boards
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: get all boards
      tags:
      - board-manager-apis
  /sys/filter:
    put:
      consumes:
      - application/json
      description: board db will be filtered using given model
      parameters:
      - description: Filter Content
        in: body
        name: FilterContent
        required: true
        schema:
          $ref: '#/definitions/model.Board'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: filtered data will be shown
      tags:
      - board-manager-apis
...
  /work/all:
    get:
      consumes:
      - application/json
      description: Get all Todos
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: Get all Todos
      tags:
      - todo-manager-apis
  /work/create:
    post:
      consumes:
      - application/json
      description: Create new Todo
      parameters:
      - description: Todo information
        in: body
        name: TodoInfo
        required: true
        schema:
          $ref: '#/definitions/model.Todo'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: Create new Todo
      tags:
      - todo-manager-apis
  /work/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a Todo with an ID specified
      parameters:
      - description: ID of the to be deleted Todo
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: Delete a Todo
      tags:
      - todo-manager-apis
  /work/search/{id}:
    get:
      consumes:
      - application/json
      description: gets todo by its ID
      parameters:
      - description: ID of needed todo
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: gets todo by its ID
      tags:
      - todo-manager-apis
  /work/updater/{id}:
    put:
      consumes:
      - application/json
      description: Update an ID specified Todo
      parameters:
      - description: ID of the to be updated Todo
        in: path
        name: id
        required: true
        type: integer
      - description: UpdateContent information
        in: body
        name: UpdateContent
        required: true
        schema:
          $ref: '#/definitions/model.Todo'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: Update an ID specified Todo
      tags:
      - todo-manager-apis
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"